pyproject.toml
[project]
name = "census-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "census>=0.8.24",
    "chromadb>=1.0.21",
    "ddgs>=9.5.5",
    "duckduckgo-search>=8.1.1",
    "flake8>=7.3.0",
    "future>=1.0.0",
    "ipykernel>=6.30.1",
    "kaleido>=1.1.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langgraph-prebuilt>=0.6.4",
    "langgraph-sdk>=0.2.6",
    "langsmith>=0.4.27",
    "numpy>=2.3.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pip>=25.2",
    "plotly>=6.3.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "ruff>=0.14.2",
    "sentence-transformers>=5.1.0",
    "spacy>=3.8.7",
    "streamlit>=1.50.0",
    "tabulate>=0.9.0",
    "tavily-python>=0.7.11",
    "tensorflow>=2.20.0",
    "trustcall>=0.0.39",
    "typing-extensions>=4.15.0",
    "us>=3.2.0",
    "wikipedia>=1.4.0",
]

[tool.pytest.ini_options]
testpaths = ["app_test_scripts"]     # Where to look for tests
pythonpath = ["."]                  # Add to Python path
python_files = ["test_*.py"]        # Test file patterns
python_classes = ["Test*"]          # Test class patterns
python_functions = ["test_*"]       # Test function patterns
